refactor(error): 예외 클래스 삭제 후 모든 오류를 IllegalArgumentException을 던지도록 변경
