+++ src/main/java/lotto/dto/WinningLottoInformation.java
@@ -1,9 +1,9 @@
 package lotto.dto;
 
+import static lotto.domain.common.ErrorMessage.BONUS_NUMBER_DUPLICATION;
+
 import java.util.HashSet;
 import lotto.domain.Lotto;
-import lotto.exception.BonusNumberDuplicationException;
-import lotto.exception.common.ErrorMessage;
 
 public record WinningLottoInformation(
         Lotto winningLotto,
@@ -11,7 +11,7 @@ public record WinningLottoInformation(
 ) {
     public static WinningLottoInformation of(Lotto winningLotto, int bonusNumber) {
         if (new HashSet<>(winningLotto.numbers()).contains(bonusNumber)) {
-            throw new BonusNumberDuplicationException(ErrorMessage.BONUS_NUMBER_DUPLICATION);
+            throw new IllegalArgumentException(BONUS_NUMBER_DUPLICATION.getMessage());
         }
 
         return new WinningLottoInformation(winningLotto, bonusNumber);
