--- src/main/java/lotto/exception/InvalidInputException.java
@@ -1,11 +1,0 @@
-package lotto.exception;
-
-import lotto.exception.common.ErrorMessage;
-import lotto.exception.common.LottoException;
-
-public class InvalidInputException extends LottoException {
-
-    public InvalidInputException(ErrorMessage message) {
-        super(message);
-    }
-}
