--- src/main/java/lotto/exception/common/LottoException.java
@@ -1,8 +1,0 @@
-package lotto.exception.common;
-
-public class LottoException extends IllegalArgumentException {
-
-    public LottoException(ErrorMessage message) {
-        super(message.getMessage());
-    }
-}
