+++ src/test/java/lotto/exception/common/ErrorMessageTest.java
@@ -1,5 +1,6 @@
 package lotto.exception.common;
 
+import lotto.domain.common.ErrorMessage;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
