+++ src/test/java/lotto/view/input/parser/BonusNumberParserTest.java
@@ -1,6 +1,6 @@
 package lotto.view.input.parser;
 
-import lotto.exception.common.ErrorMessage;
+import lotto.domain.common.ErrorMessage;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
